Moves (most) buffer-related functionality to fabtsuite_buffer
