+++ transfer/fabtsuite_fifo.c
@@ -9,6 +9,7 @@
 #include <rdma/fabric.h>
 #include <rdma/fi_endpoint.h>
 
+#include "fabtsuite_buffer.h"
 #include "fabtsuite_error.h"
 #include "fabtsuite_fifo.h"
 #include "fabtsuite_types.h"
